8514/A compatible changes of the day (March 6th, 2025)
1. Follow the Mach32 manual more closely regarding vblank support.
2. The subsystem status now takes account of the other bits more accurately.
3. The Mach32 PCI, when used with the ATI 68860 ramdac, has its own bpp's when in accelerator mode, separate from the VGA compatible side, so fix this accordingly.
4. Reset the vram when a mapping change occurs, should clear the messups in the ATI Mach8/32 accel video mode tests.

+++ src/include/86box/vid_8514a.h
@@ -213,7 +213,7 @@ typedef struct ibm8514_t {
     int     vdisp2;
     int     disp_cntl;
     int     interlace;
-    uint8_t subsys_cntl;
+    uint16_t subsys_cntl;
     uint8_t subsys_stat;
 
     atomic_int fifo_idx;
